Changing course to wrap AxisArray with FlowSample
I realised I was just rewriting the implementation of AxisArray. This
commit introduces the tests for a new FlowSample struct which is a
simple wrapper for AxisArray with some params attached.

AxisArray performance is supposed to be good, the interface doesn't
have to change, and they can easily be converted to matrices as needed
by GigaSOM.